requirements.txt
--find-links https://github.com/openxla/stablehlo/releases/expanded_assets/dev-wheels 
stablehlo
cython==3.0.11
z3-solver>=4.12.0
hydra-core
omegaconf